[DF-11317] Fixed review issues
Former-commit-id: 48e4c3541d39fcd26c8efea0b69d06a8b53a488a

+++ Sources/Internal/Render/Image/ImageConvert.h
@@ -569,7 +569,7 @@ struct PackRGBA16161616
 
 struct UnpackRGBA32323232
 {
-    inline void operator()(const RGBA32323232* input, uint32& r, uint32& g, uint32& b, uint32& a)
+    inline void operator()(const RGBA32323232* input, uint64& r, uint64& g, uint64& b, uint64& a)
     {
         r = input->r;
         g = input->g;
@@ -580,12 +580,12 @@ struct UnpackRGBA32323232
 
 struct PackRGBA32323232
 {
-    inline void operator()(uint32& r, uint32& g, uint32& b, uint32& a, RGBA32323232* out)
+    inline void operator()(uint64& r, uint64& g, uint64& b, uint64& a, RGBA32323232* out)
     {
-        out->r = r;
-        out->g = g;
-        out->b = b;
-        out->a = a;
+        out->r = static_cast<uint32>(r);
+        out->g = static_cast<uint32>(g);
+        out->b = static_cast<uint32>(b);
+        out->a = static_cast<uint32>(a);
     }
 };
